mm: compaction: fix return value of capture_free_page()
Commit ef6c5be6 ("fix incorrect NR_FREE_PAGES accounting (appears like
memory leak)") fixes a NR_FREE_PAGE accounting leak but missed the return
value which was also missed by this reviewer until today.  That return
value is used by compaction when adding pages to a list of isolated free
pages and without this follow-up fix, there is a risk of free list
corruption.

Signed-off-by: Mel Gorman <mgorman@suse.de>
Cc: Dave Hansen <dave@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>