parisc: fix race conditions flushing user cache pages
There are two issues addressed by this patch:

1) When we flush user instruction and data pages using the kernel
tmpalias region, we need to ensure that the local TLB entry for the
tmpalias page is never replaced with a different mapping. Previously, we
purged the entry before and after the cache flush loop. Although
preemption was disabled, it seemed possible that the value might change
during interrupt processing. The patch removes the purge and disables
interrupts during the initial TLB entry purge and cache flush.

2) In a number of places, we flush the TLB for the page and then flush
the page. We disabled preemption around the flush. This change disables
preemption around both the TLB and cache flushes as it seemed the effect
of the purge might be lost.

Without this change, I saw four random segmentation faults in about 1.5
days of intensive package building last weekend. With the change, I
haven't seen a single random segmentation fault in about one week of
building Debian packages on 4-way rp3440. So, there is a significant
improvement in system stability.

Signed-off-by: John David Anglin <dave.anglin@bell.net>
Signed-off-by: Helge Deller <deller@gmx.de>