drm/i915: refuse to submit more batchbuffers from guilty context
If context has recently submitted a faulty batchbuffers guilty of
gpu hang and decides to keep submitting more crap, ban it permanently.

Note: This is just a refinment of the existing policy to declare the
gpu wedged when it's hanging too often. We have that in place to
increase the odds of a system surviving gpu hang storms (and so
getting a useful bug report with the error state).

With this patch here we reduce the risk to take down unrelated
applications, since especially modern compositors tend to die when
mesa can't submit batches any more. So for the somewhat common case of
a specific gl workload causing a gpu hang flurry we should now have
less grumpy users since we don't take down their entire deskopt
enviroment any more.

v2: Store guilty ban status bool in gpu_error instead of pointers
    that might become danling before hang is declared.

v3: Use return value for banned status instead of stashing state
    into gpu_error (Chris Wilson)

Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
[danvet: Added note to commit message to explain why we want this,
which addresses a concern Ben raised.]
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>