ocfs2: fix readonly issue in ocfs2_unlink()
While deleting a file with ocfs2_unlink(), there is a bug in this
function.  This bug will result in filesystem read-only.

After calling ocfs2_orphan_add(), the file which will be deleted is added
into orphan dir.  If ocfs2_delete_entry() fails, the file still exists in
the parent dir.  And this scenario introduces a conflict of metadata.

If a file is added into orphan dir, when we put inode of the file with
iput(), the inode i_flags is setted (~OCFS2_VALID_FL) in
ocfs2_remove_inode(), and then write back to disk.

But as previously mentioned, the file still exists in the parent dir.  On
other nodes, the file can be still accessed.  When first read the file
with ocfs2_read_blocks() from disk, It will check and avalidate inode
using ocfs2_validate_inode_block().  So File system will be readonly
because the inode is invalid.  In other words, the inode i_flags has been
set (~OCFS2_VALID_FL).

Signed-off-by: Younger Liu <younger.liu@huawei.com>
Signed-off-by: Jensen <shencanquan@huawei.com>
Cc: Jie Liu <jeff.liu@oracle.com>
Cc: Joel Becker <jlbec@evilplan.org>
Cc: Mark Fasheh <mfasheh@suse.com>
Cc: Sunil Mushran <sunil.mushran@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>