ARM: dts: am335x-tx48: remove pinctrl from usb0_phy
The pins used for VBUSEN and OC in host mode are also used for the 2nd
CAN interface. When the USB0 interface is used in device mode, the
pins can still be used for the CAN interface if they are not claimed
by the USB PHY driver. Thus move the pinctrl reference to the
reg_usbotg which will only claim them for host mode.