net: fec: don't skip check of num-rx-queues, when num-tx-queues is out of range
There is no need to return from fec_enet_get_queue_num(), skipping the
check for num-rx-queues, when the value of num-tx-queues specified in
DT is invalid.
Furthermore print the name of the DT property set by the user, rather
than the driver's internal variable name in the warning message.