dt-bindings: rs485: add support of delays in nanoseconds
Add rs485-rts-delay-ns (in nanoseconds) property.
This extends the range of delay values for user to microseconds and
nanoseconds.

The code defines the rs485 delay in milliseconds.
With fast data rates (1 Mbps or more), 1 ms become quite long. Users need
to be able to set shorter delays than 1 ms in order not to slow down the
channel unnecessarily. This enables faster communication channels without
changing the baudrate.

Signed-off-by: Erwan Le Ray <erwan.leray@foss.st.com>