dts: tx8m: fix DRAM voltage setting
The BD71847 regulator used on the TX8M modules has a 2.2K/449R voltage
divider in the feedback path of the BUCK6 (BUCK8 in datasheet)
regulator that supplies the DDR3 SDRAM voltage. The default voltage of
this regulator is 1.1V that is adjusted to 1.35V via the voltage
divider. Thus setting the nominal voltage to 1.1V will result in the
proper 1.35V for the DDR3 SDRAM.