nvmem: stm32: detect bsec pta presence for stm32mp15
On STM32MP15, the SMC backend is optional when OP-TEE is used;
the PTA BSEC should be used as it is done on STM32MP13 platform,
but the BSEC SMC can be also used: it is a legacy mode in OP-TEE,
not recommended but used in previous OP-TEE firmware.

The presence of OP-TEE is dynamically detected in STM32MP15x device tree
and the supported NVMEM backend is dynamically detected:
- PTA with stm32_bsec_pta_find
- SMC with stm32_bsec_check

Without PTA and SMC detection, the probe is defered for STM32MP15 device.

On STM32MP13x platform, only the PTA is supported with cfg->ta = true
and this detection is skipped.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I59210046e368cfc22bd3cca2afe1653674f8ece8
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/266103
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>