media: stm32: dcmipp: add STM32MP25x platform support
On the STM32MP25x platform, the DCMIPP embeds 3 pipes
instead of a single pipe on the previous STM32MP13x.
In addition to the byte pipe, the STM32MP25x has two
pixel based pipes (main & aux).  The main is made of
an ISP stage followed by a post-processing stage, while
aux pipe has only a post-processing stage.  The aux
pipe input can either be straight from the input or
coming from the output of the ISP stage of the main pipe.

In addition to that, the DCMIPP can receive input from
either a parallel interface or from a CSI2 interface via
the csi2host driver.

Change-Id: I17e5c30e86c627c618acf66c6baf8c7aad481c8f
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Signed-off-by: Hugues Fruchet <hugues.fruchet@foss.st.com>