ARM: dts: stm32: add adc internal channels to stm32mp15
Add STM32 ADC2 internal channels VREFINT and VDDCORE to STM32MP15x SoCs.
VBAT internal channel is not defined by default in SoC DT, and
has be defined in board DT when needed, instead. This avoids unwanted
current consumption on battery, when ADC conversions are performed
on any other channels.

The internal channels are defined in STM32MP15 SoC DT according to
generic channel bindings. The STM32 driver does not support a mixed use
of legacy and generic channels. When generic channels are defined,
legacy channel are ignored. This involves that the board device trees
using legacy bindings for ADC2, have to be reworked.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/283303
Signed-off-by: Olivier Moysan <olivier.moysan@foss.st.com>
Change-Id: If6a9c088451d54cfed23f3aa4cffe813c264f99e