hwrng: stm32 - rework RNG configuration sequence
This change introduces a change in the application of the
configuration for the RNG at cold boot and during low-power
sequences.

Configuration should be applicable when the conditioning soft
reset is supported. Remove the dependency to the clock error
detection for this part as there is no link between the two.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: Ib12b5a29eebd6dee656bb980af8de21cd4e96e06
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/261072
Tested-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>