iio: adc: stm32: add support for vddcpu and vddgpu on stm32mp25
Add support for vddcpu and vddgpu internal channels on stm32mp25.
These channels are enabled by using ADC option register.
Add VDDGPU to the list of supported channels in this driver and made
available on stm32mp25. So update sampling information for all variants
(with zero'ed data when it's not available).

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: Ibadafff744fea7aa5869d52554fddf2a8b54433b
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/270315
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Olivier MOYSAN <olivier.moysan@foss.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>