dt-bindings: remoteproc: add the vector table sysconf registers
Add the definition of the st,syscfg-nsvtor pointing to the
register which store the non secure vector table pointer
(the trustzone is not enable).

Information is needed on the stm32mp257 SoC when the firmware is
loaded by the linux remoteproc driver. Else, by default, the
boot address is the MCU RETRAM base address.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: I46d06735fca279f9dc607cda2ee0b12af4efde2d
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/269210
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>