media: stm32-dcmipp: add support of statistics capture video device
Add support of a video device (type META capture) after the ISP subdev
on main pipe of the DCMIPP.
Purpose is to be able to capture statistics out of the main pipe and give
those information to the application so that it can then control the
pipeline.

For the time being statistic window configuration is not possible
and kept to the full frame.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>