remoteproc: stm32: clean-up the management of the hold boot by smc call
There are two ways to manage the Cortex-M4 hold boot:
- by Linux thanks to a sys config controller
- by the secure context when the hold boot is protected.
Since the introduction of the scmi server, the use of the smc call
is deprecated.If the trust zone is activated, the management of the
hold boot will be done by the secure context thanks to a smci reset
controller.

This patch cleans-up the code related to the smc call.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: I6d2d786260ee014c43f019c5ac5ca9de535d6093