arm64: dts: st: stm32: add RIFSC as a system bus for STM32MP25x boards
The STM32 System Bus is an internal bus on which devices are connected.
RIFSC is a peripheral overseeing the firewall bus that configures
and control access to the peripherals connected on it.

For more information on which peripheral is securable, please read
the STM32MP25 reference manual.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I7414a92082fa5c16c3d13887b67957bbccadd4ce
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/301993
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>
Reviewed-by: Gabriel FERNANDEZ <gabriel.fernandez@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Lionel DEBIEVE <lionel.debieve@foss.st.com>