remoteproc: stm32_rproc: Get resource table info from nvmem for the stm32mp25
For the stm32mp25 the nvmem framework is used to get resource table
information. for the stm32MP15 the syscon is used for compatibility
with the legacy
Use a function associated to the compatible to dispatch the method to
retrieve the information.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: Ia484a0ce9caaf2d50ddc9aef29bb416f9dd1967c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/326925
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>