pinctrl: stm32: instantiate one irq domain per controller
The current code instantiates one irq domain per GPIO bank.
This is currently not needed as only one GPIO, among the GPIO with
same position in each bank, can be routed to the IRQ controller by
the mux in the EXTI (for MPU) or in the st,stm32-syscfg (for MCU).

The driver is going to support multiple EXTI irq controllers and
this will further increase the number of irq domains to a total of
the number of banks multiplied by the number of EXTI.

Modify the driver to instantiate a single irq domain used by all
the GPIO banks of the controller.
Track per bank the virq used by the GPIO, if any.
While there, manage the irq domain to safely remove it on probe
error.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I23c4e676146289fe957970986a3ef6907740ba10
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/329529
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>