irqchip/rpmsg: fix IRQ domain remove
Freeing an IRQ while the IRQ domain has been removed can cause a crash.
In the current implementation, the rpmsg device can remove the IRQ domain
before the IRQ consumers free IRQs. The reason for this is that there is
a devlink between the platform device and the IRQ consumer devices, but
no devlink between the rpmsg device and the IRQ consumer devices.

To fix this, we need to create/remove the IRQ domain devlink at the
platform level.

However, a drawback of this implementation is that we have to allocate
the domain before knowing the number of IRQs to map.
To ensure that the size of the domain matches the number of IRQ banks
to support, a test is added.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: Ib019670981ba13f08d77ff37e201f739da524557
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/331500
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Michel JAOUEN <michel.jaouen@st.com>