media: dcmipp: addition of ISP parameter video device
Introduce a output video device which, placed in front of the
isp subdev, is in charge of handling ISP parameters given by
the application.  This puts the foundation of the mechanism
with main ISP controls.

Once pushed, the sequence field of the v4l2_buffer structure
should be updated to inform the application the frame into
which this will be taken into account.
(TODO - not yet done)

Change-Id: I3645a573e462134d8ad98c53bd303458b3d69d56
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/353229
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>
Reviewed-by: Hugues FRUCHET <hugues.fruchet@st.com>
Domain-Review: Philippe CORNU <philippe.cornu@foss.st.com>