media: stm32: dcmipp: correct PIPEDIFF handling within dcmipp-isp
A unique stream is getting out of the ISP can go into both the pixelproc
subdev of main pipe and aux pipe. A unique pad on isp subdev can thus be
used for that purpose. Connect the aux pixelproc to the pad[1] of the
isp. Since that pad can have 2 remotes pad, implement a dedicated function
to discover if the link to the aux pixelproc is enabled or not.

Change-Id: Ic84bc19793fc6042fcdcb6d491a242a591177e63
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/352412
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>
Domain-Review: Philippe CORNU <philippe.cornu@foss.st.com>