dt-bindings: dma: stm32-dma3: introduce lli-bus-interface property
STM32 DMA3 has up to two master ports. It always has a port 0, and could
have a port 1. Depending on hardware configuration, ports can be AHB or
AXI. Depending on where linked-list items are allocated, or depending of
the port 0 usage, the user may want to use the second port if there is one.
lli-bus-interface property can be set to 0, to select port 0, or set to 1,
to select port 1. If the property is not defined, port 0 is selected by
default.

Change-Id: I6b74755c5a2a2251af29916382d8033a4a814750
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/370082
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>