usb: dwc3: Fix for xhci-reset stuck issue with dwc3
Patch provided by Synopsys with following comment :
The change we have here is to keep both the GUSB3PIPECTL.SUSPENDENABLE
and GUSB2PHYCFG.SUSPENDUSB20 bits clear when switching mode, only to
restore them after the controller initialization completes.
For host mode, the host controller initialization described in xHCI spec
section 4.2 describe the last step of the initialization is to turn on
the Run/Stop bit. This is what we tried with drd_susphy_v2. This patch
will resolve this issue. The change touches some xhci files, not the
core files, and it will be available in linux mainline in the future

Change-Id: I6b08f16d6fc2d660c56d0ddea4f67c13e844f1c3
Signed-off-by: Pankaj Dev <pankaj.dev@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/370114
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Domain-Review: Fabrice GASNIER <fabrice.gasnier@foss.st.com>