pinctrl: stm32: Manage irq affinity settings
Trying to set the affinity of the interrupt associated to stm32
pinctrl results in a write error because stmf32 pinctrl was not
managing the set of the irq smp affinity. It is resolved by
adding the .irq_set_affinity to the stm32_gpio_irq_chip struct and
fill it with a function to manage affinity.

Change-Id: I32cb2ca38936f114874abc404cab76bc64f93c14
Signed-off-by: Cheick TRAORE <cheick.traore-ext@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/368513
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>