mfd: stmfx: Manage irq affinity settings
Trying to set the affinity of the interrupt associated to STMFx
results in a write error cause the STMFx was not managing the set
of the irq smp_affinity. It is resolved by adding the
.irq_set_affinity to the stmfx_irq_chip struct and fill it
with a function to manage affinity. The function return an
error explaining to set the affinity of the gpio interrupt
provider to change stmfx smp_affinity.

Change-Id: I56764ae5de695522670e6aa2e81ea53e54be5990
Signed-off-by: Cheick TRAORE <cheick.traore-ext@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/360860
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>