pinctrl: stmfx: Manage irq affinity settings
Trying to set the affinity of the interrupt associated to stmfx
pinctrl results in a write error because stmfx pinctrl was not
managing the set of the irq smp affinity. It is resolved by
adding the .irq_set_affinity to the stmfx pinctrl irq_chip struct and
fill it with a function to manage affinity. The function return an
error explaining to set the affinity of the gpio interrupt
provider to change stmfx pinctrl smp_affinity.

Change-Id: Icbec61a46d246cfa87e759c4e34acc2d45508768
Signed-off-by: Cheick TRAORE <cheick.traore-ext@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/368509
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>