ARM: dts: stm32: fix irq type of touchscreen on stm32mp157 EV1 boards
The touchscreen driver goodix_ts ignores the interrupt type set in
the DT; it queries the touchscreen firmware for the interrupt type
to be requested.
When the touchscreen device is unbound, the interrupt keeps the
type already requested.
On a following bind of the same driver, the interrupt framework
will find that the interrupt is already registered and will check
if its interrupt type matches the DT; in case of type mismatch the
bind will fail.

The touchscreen firmware can be update during the product life,
thus the interrupt type can change.

Use IRQ_TYPE_NONE in the DT, as it works as a wildcard and matches
any interrupt type requested by the firmware.

Current touchscreen firmware in stm32mp157 EV1 require falling
edge type, that mismatches the type in DT and prevents re-binding
the device.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I824c84bf0295feee111a66ff0409a9528ce2f8bd
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/373137
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>