imx8m/imx93: fix ethernet link up/down bug with SMSC PHY due to incorrect PHY refclock handling
The SMSC LAN87xx PHY requires the RMII refclk to be running before
RESET is deasserted and needs to be reset whenever the refclk has been
disabled and re-enabled.

With the current execution flow this cannot be guaranteed.

Add an optional clock to the mdiobus driver that is requested and
enabled before doing the reset procedure to achieve this.